Improving messgaes when creating links

+++ link.bash
@@ -2,6 +2,9 @@
 # creates links in the parent directory to all dotfiles in this directory.
 # HAS to be run in this directory
 
+set -e
+set -u
+
 # e.g.Turns ".cabal/config" into "../"
 function to_breadcrumbs() {
     echo "$1" | sed -r -e 's@[^/]@@g' -e 's@/@../@g'
@@ -13,12 +16,22 @@ COPIED_FILES=""
 chmod u+x bin/*
 
 for file in $LINKED_FILES; do
-    mkdir -p ../$(dirname $file)
-    ln -s $(to_breadcrumbs $file)dotfiles/$file ../$file;
+    if [[ ! -e ../$file ]]; then
+        echo LINKING $file
+        mkdir -p ../$(dirname $file)
+        ln -s $(to_breadcrumbs $file)dotfiles/$file ../$file
+    else
+        echo LINK $file already exists
+    fi
 done
 
 for file in $COPIED_FILES; do
-    cp $file ../$file;
+    if [[ ! -e ../$file ]]; then
+        echo COPYING $file
+        cp $file ../$file;
+    else
+        echo COPY $file already exists
+    fi
 done
 
 # Fix warning from ghci complaining it is writable
